Refactoing #2 and now MouseButtonEvent and KeyboardKeyEvent contains FarLightKeyKode.

+++ FarLight/src/FarLight/EventSystem/KeyboardEvents/KeyboardKeyPressedEvent.cpp
@@ -10,7 +10,7 @@ namespace FarLight
     std::string KeyboardKeyPressedEvent::ToString() const noexcept
     {
         std::stringstream ss;
-        ss << std::boolalpha << GetName() << ": " << "(" << GetKeyCode() << "). " << "Is repeated: " << IsRepeated() << ".";
+        ss << std::boolalpha << GetName() << ": " << "(" << static_cast<int>(GetKeyCode()) << "). " << "Is repeated: " << IsRepeated() << ".";
         return ss.str();
     }
 }
